DoAnPtit_Xray2CT/requirements.txt
absl-py
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
aiostream
annotated-types
antlr4-python3-runtime
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
async-timeout
attrs
babel
backoff
beautifulsoup4
bleach
boto3
botocore
cachetools
certifi
cffi
charset-normalizer
click
cmocean
colorcet
comm
contourpy
cryptography
cycler
debugpy
decorator
defusedxml
Deprecated
diffdrr
diffusers
docarray
docker
einops
einops-exts
exceptiongroup
executing
fastapi
fastcore
fastjsonschema
fastremap
filelock
fill_voids
fonttools
fqdn
frozenlist
fsspec
ftfy
gdown
generative
google-auth
google-auth-oauthlib
greenlet
grpcio
grpcio-health-checking
grpcio-reflection
h11
hf-xet
httpcore
httptools
httpx
huggingface-hub
humanize
hydra-core
idna
imageio
importlib_metadata
importlib_resources
ipykernel
ipython
ipywidgets
isoduration
jedi
Jinja2
jmespath
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_proxy
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kiwisolver
kornia
kornia_rs
lazy_loader
lightning
lightning-cloud
lightning-utilities
lightning_sdk
litdata
litserve
lpips
lungmask
lz4
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
meshio
mistune
monai
monai-generative
more-itertools
mpmath
msgpack
multidict
mypy_extensions
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nibabel
notebook_shim
numpy
oauthlib
omegaconf
open_clip_torch
opencv-python
orjson
overrides
packaging
pandas
pandocfilters
params
parso
pexpect
pillow
platformdirs
pooch
prometheus_client
prompt_toolkit
propcache
protobuf
psutil
ptyprocess
pure_eval
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydicom
Pygments
PyJWT
pylidc
pyparsing
PySocks
python-dateutil
python-dotenv
python-json-logger
python-multipart
pytorch-lightning
pytz
pyvista
PyYAML
pyzmq
referencing
regex
requests
requests-oauthlib
rfc3339-validator
rfc3986-validator
rich
rotary-embedding-torch
rpds-py
rsa
s3transfer
safetensors
scikit-image
scikit-learn
scikit-video
scipy
scooby
seaborn
Send2Trash
shellingham
simpervisor
simple-term-menu
simpleitk
six
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
sympy
tensorboard
tensorboard-data-server
terminado
threadpoolctl
tifffile
timm
tinycss2
tokenizers
tomli
torch
torchio
torchmetrics
torchvision
tornado
tqdm
traitlets
trame
trame-client
trame-common
trame-server
trame-vtk
trame-vuetify
transformers
triton
typer
types-python-dateutil
types-requests
typing-inspect
typing-inspection
typing_extensions
tzdata
uri-template
urllib3
uvicorn
uvloop
vtk
watchfiles
wcwidth
webcolors
webencodings
websocket-client
websockets
Werkzeug
wget
widgetsnbextension
wrapt
wslink
yarl
zipp